Working thread pool implementation
